# docs/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QUEST API",
    "version": "0.0.0-dev",
    "description": "Generated from the zod contracts in @quest/types (ADR-012). Conventions: docs/api/API_CONVENTIONS.md. Regenerate with `pnpm --filter @quest/api openapi:generate`."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "local"
    }
  ],
...
  "paths": {
...
    "/v1/interests": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Interest catalogue",
        "operationId": "get-interests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string",
                            "pattern": "^[a-z][a-z0-9-]{1,39}$"
                          },
                          "label": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string"
                          },
                          "sortOrder": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "key",
                          "label",
                          "category",
                          "sortOrder"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [],
        "x-quest-permissions": [],
        "x-quest-account-states": []
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "The signed-in account",
        "operationId": "get-me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "email": {
                      "type": "string"
                    },
                    "emailVerified": {
                      "type": "boolean"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING_VERIFICATION",
                        "ACTIVE",
                        "DEACTIVATED",
                        "SUSPENDED",
                        "DELETION_REQUESTED",
                        "DELETED"
                      ]
                    },
                    "ageBand": {
                      "type": "string",
                      "enum": [
                        "TEEN_13_15",
                        "TEEN_16_17",
                        "ADULT"
                      ]
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "USER",
                          "SUPER_ADMIN",
                          "TRUST_SAFETY_LEAD",
                          "MODERATOR",
                          "SUPPORT",
                          "ANALYST",
                          "READ_ONLY"
                        ]
                      }
                    },
                    "onboarding": {
                      "type": "object",
                      "properties": {
                        "completed": {
                          "type": "boolean"
                        },
                        "nextStep": {
                          "type": "string",
                          "enum": [
                            "VERIFY_EMAIL",
                            "PROFILE",
                            "INTERESTS",
                            "DONE"
                          ]
                        }
                      },
                      "required": [
                        "completed",
                        "nextStep"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "deletionScheduledFor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "accountId",
                    "email",
                    "emailVerified",
                    "state",
                    "ageBand",
                    "roles",
                    "onboarding",
                    "createdAt",
                    "deletionScheduledFor"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      }
    },
    "/v1/me/blocks": {
      "get": {
        "tags": [
          "privacy"
        ],
        "summary": "Blocked accounts",
        "operationId": "get-me-blocks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "accountId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "username": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "blockedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "accountId",
                          "username",
                          "blockedAt"
                        ]
                      }
                    },
                    "pageInfo": {
                      "type": "object",
                      "properties": {
                        "nextCursor": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "nextCursor",
                        "hasMore"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pageInfo"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_BLOCKS"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ]
      },
      "post": {
        "tags": [
          "privacy"
        ],
        "summary": "Block an account",
        "operationId": "post-me-blocks",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_BLOCKS"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "x-quest-rate-limit": {
          "limit": 30,
          "windowSeconds": 60
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "required": [
                  "accountId"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/me/blocks/{accountId}": {
      "delete": {
        "tags": [
          "privacy"
        ],
        "summary": "Unblock an account",
        "operationId": "delete-me-blocks-by-accountId",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_BLOCKS"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ]
      }
    },
...
    "/v1/me/interests": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Selected interests",
        "operationId": "get-me-interests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "interestKeys": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "pattern": "^[a-z][a-z0-9-]{1,39}$"
                      }
                    }
                  },
                  "required": [
                    "interestKeys"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_PROFILE"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Replace selected interests",
        "operationId": "put-me-interests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "interestKeys": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "pattern": "^[a-z][a-z0-9-]{1,39}$"
                      }
                    }
                  },
                  "required": [
                    "interestKeys"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_PROFILE"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interestKeys": {
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9-]{1,39}$"
                    }
                  }
                },
                "required": [
                  "interestKeys"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/me/onboarding": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Onboarding status",
        "operationId": "get-me-onboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "completed": {
                      "type": "boolean"
                    },
                    "nextStep": {
                      "type": "string",
                      "enum": [
                        "VERIFY_EMAIL",
                        "PROFILE",
                        "INTERESTS",
                        "DONE"
                      ]
                    },
                    "missing": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "EMAIL_VERIFIED",
                          "USERNAME",
                          "DISPLAY_NAME",
                          "INTERESTS"
                        ]
                      }
                    }
                  },
                  "required": [
                    "completed",
                    "nextStep",
                    "missing"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_PROFILE"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ]
      }
    },
...
    "/v1/profiles/username-availability": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Username availability",
        "operationId": "get-profiles-username-availability",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "available": {
                      "type": "boolean"
                    },
                    "reason": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "RESERVED",
                            "TAKEN",
                            "INVALID"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "username",
                    "available",
                    "reason"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [],
        "x-quest-permissions": [],
        "x-quest-account-states": [],
        "x-quest-rate-limit": {
          "limit": 30,
          "windowSeconds": 60
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "PAYLOAD_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              },
              "correlationId": {
                "type": "string"
              },
              "timestamp": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "correlationId",
              "timestamp"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}